requirements.txt
pandas
python-dotenv
openpyxl
requests
pyodbc
chardet